adds openapi spec
Added a script that generates an openapi.json file in the public folder. Having it in the public folder allows us to serve it locally at localhost:1337/openapi.json so our front-end codegen can read and consume the file.

I gitignored it because we don't need/want the file in prod.